Rename some form components and add SelectVideo component for CCPC.js
